+++ pathway3/NavigationCodelab/app/src/main/java/com/example/compose/rally/RallyActivity.kt
@@ -21,22 +21,14 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Scaffold
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.navigation.NavController
-import androidx.navigation.NavGraph.Companion.findStartDestination
-import androidx.navigation.NavHostController
-import androidx.navigation.NavType
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
-import com.example.compose.rally.ui.accounts.AccountsScreen
-import com.example.compose.rally.ui.accounts.SingleAccountScreen
-import com.example.compose.rally.ui.bills.BillsScreen
+import com.example.compose.rally.ui.RallyNavHost
 import com.example.compose.rally.ui.components.RallyTabRow
-import com.example.compose.rally.ui.overview.OverviewScreen
+import com.example.compose.rally.ui.navigateSingleTop
 import com.example.compose.rally.ui.theme.RallyTheme
 
 /**
@@ -71,52 +63,8 @@ fun RallyApp() {
                 )
             }
         ) { innerPadding ->
-            NavHost(
-                navController = navController,
-                startDestination = Overview.route,
-                modifier = Modifier.padding(innerPadding)
-            ) {
-                composable(route = Overview.route) {
-                    OverviewScreen(
-                        onClickSeeAllAccounts = { navController.navigateSingleTop(Accounts.route) },
-                        onClickSeeAllBills = { navController.navigateSingleTop(Bills.route) },
-                        onAccountClick = { type ->
-                            navController.navigateToSingleAccount(type)
-                        }
-                    )
-                }
-
-                composable(route = Accounts.route) {
-                    AccountsScreen(
-                        onAccountClick = { type ->
-                            navController.navigateToSingleAccount(type)
-                        }
-                    )
-                }
-                composable(route = Bills.route) {
-                    BillsScreen()
-                }
-
-                composable(
-                    route = SingleAccount.routeWithArgs,
-                    arguments = SingleAccount.arguments,
-                    deepLinks = SingleAccount.deeplinks
-                ) { navBackStack ->
-                    val accountType = navBackStack.arguments?.getString(SingleAccount.accountTypeArg)
-                    SingleAccountScreen(accountType)
-                }
-            }
+            RallyNavHost(navController = navController, modifier = Modifier.padding(innerPadding))
         }
     }
-}
-
-private fun NavController.navigateSingleTop(route: String) = navigate(route) {
-    launchSingleTop = true
-    popUpTo(graph.findStartDestination().id) { saveState = true }
-    restoreState = true
-}
-
-private fun NavHostController.navigateToSingleAccount(type: String) {
-    navigateToSingleAccount("${SingleAccount.route}/$type")
 }
